Add unit tests project
This change lays the foundation for adding tests to the solution.

Directory structure:

First, the program was moved to the "src" subdirectory so that the
dotnet command line interface (and msbuild and OmniSharp) will not be
confused by the csproj file at the repository root.

An XUnit test project was created in the "test" subdirectory.

New tests:

Four tests were added for the code that modifies the markdown file:
- files with no template are not changed
- contributors are appended for a well formed contributors template
- duplicate contributors are not added
- contributors that differ only in case are considered duplicates

Other changes:

To support the unit tests, the AddContributorsToMarkdownFile method was
moved to a separate class and the method was modified slightly to move
the file I/O responsibility to the caller (WriteContributorsToRepo).

Also, an intermediate list was being built up that was not needed, and
was removed.